pyproject.toml
[tool.poetry]
name = "progress-django-api"
version = "0.1.0"
description = ""
authors = ["syuji-higa <syuji.higa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
psycopg2-binary = "^2.8.5"
djangorestframework = "^3.11.1"
django-cors-headers = "^3.5.0"
drf-yasg = "^1.17.1"
djangorestframework-camel-case = "^1.2.0"
django-3-jet = "^1.0.8"
django-admin-sortable = "^2.2.3"
factory_boy = "^3.0.1"
djangorestframework-api-key = "^2.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
mypy = "^0.782"
yapf = "^0.30.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
